refactor : 도메인과 관련된 객체를 하나의 package로 묶음.
