+++ src/main/java/racing/domain/CarFactory.java
@@ -1,4 +1,4 @@
-package racing;
+package racing.domain;
 
 import java.util.List;
 import java.util.stream.Collectors;
